[REF] Separado arquivo e recalculo de saldos
- Quando não considerado os lançamentos de fechamento para o
balancete
