+++ l10n_br_contabilidade/reports/__init__.py
@@ -1,2 +1,3 @@
 from . import general_ledger
 from . import common_reports
+from . import common_balance_reports
